Fixed SSAO. Introduced a new buffer that holds linear depth information. It might be a waste considering that theres an unused 16bit channel in the nomral vector buffer, also in theory the original depth buffer can be used to hold depth information in view space. That could be a good solution but for now im glad that it works.

+++ src/Glitter-Engine/implementation/Camera.cpp
@@ -7,12 +7,12 @@ Frustum Frustum::Transformed(const glm::mat4& transform)
 {
 	Frustum result;
 	
-	for (int i = 0; i < nearPlane.size(); i++) {
+	for (size_t i = 0; i < nearPlane.size(); i++) {
 		auto transformed = transform * glm::vec4(nearPlane[i], 1);
 		result.nearPlane[i] = glm::vec3(transformed / transformed.w);
 	}
 	
-	for (int i = 0; i < farPlane.size(); i++) {
+	for (size_t i = 0; i < farPlane.size(); i++) {
 		auto transformed = transform * glm::vec4(farPlane[i], 1);
 		result.farPlane[i] = glm::vec3(transformed / transformed.w);
 	}
